v1.1.0

+++ autostart.sh
@@ -22,9 +22,13 @@ usage_exit(){
 }
 
 on(){
-	sed -i "s@^WorkingDirectory=.*@WorkingDirectory=${SCRIPT_DIR}@" ${CMD}.service
-	sed -i "s@^ExecStart=.*@ExecStart=${SCRIPT_DIR}/${CMD}@" ${CMD}.service
-	sudo ln -s ${SCRIPT_DIR}\/${CMD}.service /etc/systemd/system/${CMD}.service
+  # create systemctl unit file if not exist
+  if [ ! -f ./test ]; then
+    cp autostart/autostart.service ${CMD}.service
+    sed -i "s@^WorkingDirectory=.*@WorkingDirectory=${SCRIPT_DIR}@" ${CMD}.service
+    sed -i "s@^ExecStart=.*@ExecStart=${SCRIPT_DIR}/${CMD}@" ${CMD}.service
+    sudo ln -s ${SCRIPT_DIR}\/${CMD}.service /etc/systemd/system/${CMD}.service
+  fi
 #	sudo cp ${SCRIPT_DIR}\/${CMD}.service /etc/systemd/system/${CMD}.service
 	sudo systemctl daemon-reload
 	sudo systemctl enable ${CMD}.service
